OHRM5X-1153: GDPR access -add controller for Download employee records (access record)
