+++ symfony/plugins/orangehrmMaintenancePlugin/Controller/File/AccessEmployeeFileController.php
@@ -1,0 +1,69 @@
+<?php
+/**
+ * OrangeHRM is a comprehensive Human Resource Management (HRM) System that captures
+ * all the essential functionalities required for any enterprise.
+ * Copyright (C) 2006 OrangeHRM Inc., http://www.orangehrm.com
+ *
+ * OrangeHRM is free software; you can redistribute it and/or modify it under the terms of
+ * the GNU General Public License as published by the Free Software Foundation; either
+ * version 2 of the License, or (at your option) any later version.
+ *
+ * OrangeHRM is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
+ * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License along with this program;
+ * if not, write to the Free Software Foundation, Inc., 51 Franklin Street, Fifth Floor,
+ * Boston, MA  02110-1301, USA
+ */
+
+namespace OrangeHRM\Maintenance\Controller\File;
+
+use OrangeHRM\Core\Controller\AbstractFileController;
+use OrangeHRM\Framework\Http\Request;
+use OrangeHRM\Framework\Http\Response;
+use OrangeHRM\Maintenance\Service\MaintenanceService;
+
+
+class AccessEmployeeFileController extends AbstractFileController
+{
+    /**
+     * @var MaintenanceService|null
+     */
+    protected ?MaintenanceService $maintenanceService = null;
+
+    /**
+     * @return MaintenanceService
+     */
+    public function getMaintenanceService(): MaintenanceService
+    {
+        if (!$this->maintenanceService instanceof MaintenanceService) {
+            $this->maintenanceService = new MaintenanceService();
+        }
+        return $this->maintenanceService;
+    }
+
+    /**
+     * @param Request $request
+     * @return Response
+     */
+    public function handle(Request $request): Response
+    {
+        $empNumber = $request->attributes->get('empNo');
+        $response = $this->getResponse();
+
+        if($empNumber){
+            $content = json_encode($this->getMaintenanceService()->accessEmployeeData($empNumber));
+            $this->setCommonHeadersToResponse(
+                'employee.json',
+                'application/json',
+                mb_strlen($content),
+                $response
+            );
+            $response->setContent($content);
+            return $response;
+        }
+
+        return $this->handleBadRequest();
+    }
+}
